requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
bidict==0.23.1
blinker==1.9.0
brotli==1.2.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
ConfigArgParse==1.7.1
et_xmlfile==2.0.0
fastapi==0.121.1
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
gevent==25.9.1
geventhttpclient==2.3.5
greenlet==3.2.4
h11==0.16.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
locust==2.42.3
locust-cloud==1.28.1
MarkupSafe==3.0.3
msgpack==1.1.2
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
platformdirs==4.5.0
pluggy==1.6.0
psutil==7.1.3
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-engineio==4.12.3
python-socketio==5.14.3
pytz==2025.2
pyzmq==27.1.0
requests==2.32.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
websocket-client==1.9.0
Werkzeug==3.1.3
wsproto==1.3.1
zope.event==6.1
zope.interface==8.1.1
